Move set_once_ip_extcommunity_list_evpn_soo to utils
